requirements.txt
alembic==0.8.3
dominate==2.1.16
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-Migrate==1.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
gunicorn==19.3.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.2
MarkupSafe==0.23
python-editor==0.4
python-instagram==1.3.2
requests==2.8.1
simplejson==3.8.0
six==1.10.0
virtualenv==13.1.2
Werkzeug==0.10.4
wheel==0.26.0
